Removes the DTO crap and adds PrizePool objects code
This commit removes the DTO pattern garbage. It isn't providing anything
useful in this application and is merely increasing the complexity. The
service pattern is remaining for creating, finding, and deleting
persisted entities. Updating entities will be done by directly modifying
the entities (at least for now).

This commit also adds code for the PrizePool objects.