Attempts to fix issues with extended character sets and Hibernate
Hibernate defaults to using whatever encoding the JVM is set to. You
can configure the encoding that Hibernate uses for most queries, but
you cannot configure the encoding hbm2ddl uses for creating schemas.
This commit fixes it in places where it can and adds a note where it
can't.