Fixes calculation of money pool entries (really bad implementation here)
This commit adjusts the calculation of money pool entries to be based
on the number of coins per person. That is, if I were to submit an entry
into the lotto of "4g" then I would get 4 separate entries into the
money draw, not a single entry.

The implementation in the patch is O(n). It should really be improved,
but I think that is going to require separate tracking of gold and
"entries". I'm not too keen on that idea as I really like the way the
data is tracked right now. This will require more thought and can
be fixed later.